+++ lib/PhpParser/Builder/Trait_.php
@@ -9,6 +9,7 @@ use PhpParser\Node\Stmt;
 class Trait_ extends Declaration
 {
     protected $name;
+    protected $uses = [];
     protected $properties = [];
     protected $methods = [];
 
@@ -35,6 +36,8 @@ class Trait_ extends Declaration
             $this->properties[] = $stmt;
         } else if ($stmt instanceof Stmt\ClassMethod) {
             $this->methods[] = $stmt;
+        } else if ($stmt instanceof Stmt\TraitUse) {
+            $this->uses[] = $stmt;
         } else {
             throw new \LogicException(sprintf('Unexpected node of type "%s"', $stmt->getType()));
         }
@@ -50,7 +53,7 @@ class Trait_ extends Declaration
     public function getNode() : PhpParser\Node {
         return new Stmt\Trait_(
             $this->name, [
-                'stmts' => array_merge($this->properties, $this->methods)
+                'stmts' => array_merge($this->uses, $this->properties, $this->methods)
             ], $this->attributes
         );
     }