+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Version 3.1.1-dev
 ### Fixed
 
 * Fixed syntax error on comment after brace-style namespace declaration. (#412)
+* Added support for TraitUse statements in trait builder. (#413)
 
 Version 3.1.0 (2017-07-28)
 --------------------------